Fix size_t definition

+++ include/stddef.h
@@ -1,7 +1,7 @@
 #ifndef _STDDEF_H_
 # define _STDDEF_H_
 
-typedef long size_t;
+typedef unsigned long long size_t;
 #define NULL ((void*)0)
 
 #endif	//_STDDEF_H_
